Time based frame updates; memory leaks solved

+++ src/headers/structs.h
@@ -15,7 +15,7 @@ typedef struct TerminalCanvas
 typedef struct GameVariables
 {
     bool stop;
-    uint32_t x_pos;
-    uint32_t y_pos;
+    float x_pos;
+    float y_pos;
     bool i, j;
 } GameVars_t, GameVariables_t, CanvasVariables_t;
